refactor:  MediaContentMain -> MediaContentHeader
